Move CachingHiveMetastore to metastore module
