+++ lib/trino-metastore/src/main/java/io/trino/metastore/RawHiveMetastoreFactory.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.hive.metastore;
+package io.trino.metastore;
 
 import com.google.inject.BindingAnnotation;
