+++ lib/trino-metastore/src/test/java/io/trino/metastore/cache/TestImpersonationCachingConfig.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.hive.metastore.cache;
+package io.trino.metastore.cache;
 
 import com.google.common.collect.ImmutableMap;
 import io.airlift.units.Duration;
