+++ lib/trino-metastore/src/test/java/io/trino/metastore/cache/TestReentrantBoundedExecutor.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.hive.metastore.cache;
+package io.trino.metastore.cache;
 
 import com.google.common.util.concurrent.SettableFuture;
 import org.junit.jupiter.api.Test;
