+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/metastore/CachingHiveMetastoreModule.java
@@ -19,11 +19,12 @@ import com.google.inject.Scopes;
 import com.google.inject.Singleton;
 import io.airlift.configuration.AbstractConfigurationAwareModule;
 import io.trino.metastore.HiveMetastoreFactory;
-import io.trino.plugin.hive.metastore.cache.CachingHiveMetastore;
-import io.trino.plugin.hive.metastore.cache.CachingHiveMetastoreConfig;
-import io.trino.plugin.hive.metastore.cache.ImpersonationCachingConfig;
-import io.trino.plugin.hive.metastore.cache.SharedHiveMetastoreCache;
-import io.trino.plugin.hive.metastore.cache.SharedHiveMetastoreCache.CachingHiveMetastoreFactory;
+import io.trino.metastore.RawHiveMetastoreFactory;
+import io.trino.metastore.cache.CachingHiveMetastore;
+import io.trino.metastore.cache.CachingHiveMetastoreConfig;
+import io.trino.metastore.cache.ImpersonationCachingConfig;
+import io.trino.metastore.cache.SharedHiveMetastoreCache;
+import io.trino.metastore.cache.SharedHiveMetastoreCache.CachingHiveMetastoreFactory;
 
 import java.util.Optional;
