+++ testing/trino-product-tests/src/main/java/io/trino/tests/product/deltalake/TestDeltaLakeJmx.java
@@ -35,7 +35,7 @@ public class TestDeltaLakeJmx
     {
         assertThat(onTrino().executeQuery("SHOW TABLES IN jmx.current LIKE '%name=delta%'")).containsOnly(
                 row("io.trino.filesystem.s3:name=delta,type=s3filesystemstats"),
-                row("io.trino.plugin.hive.metastore.cache:name=delta,type=cachinghivemetastore"),
+                row("io.trino.metastore.cache:name=delta,type=cachinghivemetastore"),
                 row("io.trino.plugin.hive.metastore.glue:name=delta,type=gluehivemetastore"),
                 row("io.trino.plugin.hive.metastore.glue:name=delta,type=gluemetastorestats"),
                 row("io.trino.plugin.base.metrics:catalog=delta,name=delta,type=fileformatdatasourcestats"),
@@ -48,7 +48,7 @@ public class TestDeltaLakeJmx
     {
         assertThat(onTrino().executeQuery("SHOW TABLES IN jmx.current LIKE '%name=delta%'")).containsOnly(
                 row("io.trino.filesystem.s3:name=delta,type=s3filesystemstats"),
-                row("io.trino.plugin.hive.metastore.cache:name=delta,type=cachinghivemetastore"),
+                row("io.trino.metastore.cache:name=delta,type=cachinghivemetastore"),
                 row("io.trino.plugin.hive.metastore.thrift:name=delta,type=thrifthivemetastore"),
                 row("io.trino.plugin.hive.metastore.thrift:name=delta,type=thriftmetastorestats"),
                 row("io.trino.plugin.base.metrics:catalog=delta,name=delta,type=fileformatdatasourcestats"),
